Done refactor
