+++ main.py
@@ -5,13 +5,12 @@ from fastapi.middleware.cors import CORSMiddleware
 from motor.motor_asyncio import AsyncIOMotorClient
 from decouple import config
 from db import Sport, SportSchedule, SportType, RefereeID
-from auth.auth_handler import signJWT
+from auth.auth_handler import check_password, check_user, hash_password, create_access_token
 from auth.cookie import OAuth2PasswordBearerWithCookie
 from models import RefereeIdBody, SportScheduleBody
 from utils import error_handler
 from Enum.sportStatus import SportStatus
 import bcrypt
-
 
 app = FastAPI()
 oauth2_scheme = OAuth2PasswordBearerWithCookie(tokenUrl="/login/verify")
@@ -97,55 +96,19 @@ async def add_data():
     await SportType(**sport_type_body).insert()
     await SportSchedule(**sport_schedule_body).insert()
 
-async def check_user(id_body: RefereeIdBody):
-    """Check if user exists in database return boolean"""
-    user = await RefereeID.find_one(RefereeID.username == str(id_body.username))
-    return user is not None
-
-async def check_password(id_body: RefereeIdBody):
-    """Check if password is correct return boolean"""
-    password = await RefereeID.find_one(RefereeID.username == str(id_body.username))
-    return bool(bcrypt.checkpw(id_body.password.encode("utf-8"), password.password.encode("utf-8")))
-    # return password is not None
-
-def hash_password(id_body: RefereeIdBody):
-    """Hash password"""
-    return bcrypt.hashpw(id_body.password.encode("utf-8"), bcrypt.gensalt())
-
-@app.get('/user/{username}/{password}')
-async def get_user(username, password):
-    """user for testing purpose"""
-    #TODO Delete this once finish testing
-    return await RefereeID.find_one(RefereeID.username == str(username), RefereeID.password == str(password))
-
-@app.post('/create/user')
-async def create_user(id_body: RefereeIdBody):
-    """Create user for testing purpose"""
-    data = {
-        "username": id_body.username, 
-        "password": hash_password(id_body).decode("utf-8")
-    }
-    user = RefereeID(**data)
-    await user.insert()
-    return {
-        "status": "success",
-        "message": "User created successfully",
-        "data": user
-    }
-
 @error_handler
 @app.post('/login/verify', status_code=201)
 async def login(response: Response, id_body: RefereeIdBody):
     """
     Receive username and password from body and check if user exists in database and password is correct
-    then set cookie with the access token, after that redirect to homepage
+    then set cookie with the access token that expires in 24 days, after that redirect to homepage
     """
     if await check_user(id_body) and await check_password(id_body):
-        access_token = signJWT(id_body.username)
+        access_token = create_access_token(id_body.username)
         response = RedirectResponse(
             url='/test/verify', status_code=302)  # change URL to homepage
         response.set_cookie(key="access_token",
-                            value=f"Bearer {access_token}", httponly=True)
+                            value=f"Bearer {access_token}", httponly=True, expires=2073600)
         return response
     raise HTTPException(401, "Invalid credentials")
